Simplify page layout
This adds a pages directory for the main pages so that they don't have
to live in the top level of the directory structure of the project.
Permalinks will ensure they end up at the correct path.

Also removed the about layout since this was unnecessary. The about page
now uses the default layout.